requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pypdf>=3.17.4
openai>=1.0.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
aiofiles>=23.0.0
pydantic>=2.0.0
pillow>=10.0.0
pytesseract>=0.3.10
alibabacloud-ocr-api20210707>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
